requirements.txt
langchain
langchain_community
langchain_core
langchain_openai
tiktoken
duckduckgo-search
python-dotenv
streamlit
